requirements.txt
flask>=3.0
youtube-transcript-api>=0.6
pywebview>=5.0
